requirements.txt
streamlit
numpy
pillow
googletrans==4.0.0-rc1
gtts
gdown
tensorflow-cpu==2.12.0